@@ -3,14 +3,13 @@ description: Game-based Investments
 
 publish_to: "none"
 
-version: 3.1.0+43
+version: 3.1.2+44
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   cached_network_image: ^3.1.0
-  camera: ^0.8.1+7
   cloud_firestore: ^2.4.0
   confetti: ^0.5.4+1
   connectivity_plus: ^1.1.0
@@ -35,7 +34,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_facebook_sdk: ^1.0.0
-  flutter_html: ^2.1.1
+  # flutter_html: ^2.1.1
   flutter_html_to_pdf: ^0.6.1
   flutter_image_compress: ^1.1.0
   flutter_localizations:
@@ -54,6 +53,7 @@ dependencies:
   intl: ^0.17.0
   logger: ^1.0.0
   lottie: ^1.1.0
+  webengage_flutter: ^1.0.3
   mixpanel_flutter: ^1.3.1
   open_file: ^3.2.1
   package_info: ^2.0.0
@@ -64,7 +64,7 @@ dependencies:
   property_change_notifier: ^0.3.0
   provider: ^5.0.0
   razorpay_flutter: ^1.2.7
-  rive: ^0.7.22
+  rive: 0.7.28
   share_plus: ^2.1.4
   shared_preferences: ^2.0.6
   shimmer_animation: ^2.1.0+1
@@ -73,12 +73,13 @@ dependencies:
   synchronized: ^3.0.0
   url_launcher: ^6.0.9
   video_player: ^2.1.12
-  webview_flutter: ^2.0.13
+  webview_flutter: ^3.0.0
   flutter_rating_bar: ^4.0.0
   in_app_review: ^2.0.3
-
-
-
+  scratcher: ^2.1.0
+  timelines: ^0.1.0
+  sliding_up_panel: ^2.0.0+1
+  app_install_date: ^0.1.1
 
 
 dev_dependencies:
@@ -96,6 +97,7 @@ flutter:
     - resources/augmont/fello-augmont-invoice.html
     - resources/fello-email-verification.html
     - resources/public.key
+    - resources/freshdesk.html
 
     - assets/images/
     - assets/images/icons/
