@@ -4,18 +4,18 @@ import 'package:flutter/material.dart';
 enum Flavor { DEV, QA, PROD }
 
 class FlavorValues {
-  FlavorValues(
-      {@required this.awsIciciStage,
-      @required this.awsAugmontStage,
-      @required this.freshchatStage,
-      @required this.signzyStage,
-      @required this.signzyPanStage,
-      @required this.razorpayStage,
-      @required this.baseUriAsia,
-      @required this.baseUriUS,
-      @required this.dynamicLinkPrefix,
-       @required this.mixpanelToken,
-      });
+  FlavorValues({
+    @required this.awsIciciStage,
+    @required this.awsAugmontStage,
+    @required this.freshchatStage,
+    @required this.signzyStage,
+    @required this.signzyPanStage,
+    @required this.razorpayStage,
+    @required this.baseUriAsia,
+    @required this.baseUriUS,
+    @required this.dynamicLinkPrefix,
+    @required this.mixpanelToken,
+  });
 
   final AWSIciciStage awsIciciStage;
   final FreshchatStage freshchatStage;
