@@ -2,13 +2,16 @@ class VerifyPanResponseModel {
   bool flag;
   String upstreamName;
   String message;
+  String gtId;
 
-  VerifyPanResponseModel({this.flag, this.upstreamName, this.message});
+  VerifyPanResponseModel(
+      {this.flag, this.upstreamName, this.message, this.gtId});
 
   VerifyPanResponseModel.fromJson(Map<String, dynamic> json) {
     flag = json['flag'];
     upstreamName = json['upstreamName'];
     message = json['message'];
+    gtId = json['gtId'];
   }
 
   Map<String, dynamic> toJson() {
@@ -16,6 +19,7 @@ class VerifyPanResponseModel {
     data['flag'] = this.flag;
     data['upstreamName'] = this.upstreamName;
     data['message'] = this.message;
+    data['gtId'] = this.gtId;
     return data;
   }
 
@@ -24,17 +28,19 @@ class VerifyPanResponseModel {
       'flag': flag,
       'upstreamName': upstreamName,
       'message': message,
+      'gtId': gtId,
     };
   }
 
   factory VerifyPanResponseModel.fromMap(Map<String, dynamic> map) {
     return VerifyPanResponseModel(
-      flag: map['flag'] ?? false,
-      upstreamName: map['upstreamName'] ?? '',
-      message: map['message'] ?? '',
-    );
+        flag: map['flag'] ?? false,
+        upstreamName: map['upstreamName'] ?? '',
+        message: map['message'] ?? '',
+        gtId: map['gtId'] ?? '');
   }
 
   @override
-  String toString() => 'VerifyPanResponseModel(flag: $flag, upstreamName: $upstreamName, message: $message)';
+  String toString() =>
+      'VerifyPanResponseModel(flag: $flag, upstreamName: $upstreamName, message: $message , gtId: $gtId)';
 }
