@@ -10,9 +10,10 @@ import 'package:felloapp/core/model/user_transaction_model.dart';
 import 'package:felloapp/core/ops/db_ops.dart';
 import 'package:felloapp/core/ops/razorpay_ops.dart';
 import 'package:felloapp/core/repository/investment_actions_repo.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
 import 'package:felloapp/core/service/api_service.dart';
 import 'package:felloapp/core/service/augmont_invoice_service.dart';
-import 'package:felloapp/core/service/mixpanel_service.dart';
+import 'package:felloapp/core/service/golden_ticket_service.dart';
 import 'package:felloapp/core/service/transaction_service.dart';
 import 'package:felloapp/core/service/user_coin_service.dart';
 import 'package:felloapp/core/service/user_service.dart';
@@ -23,7 +24,7 @@ import 'package:felloapp/util/fail_types.dart';
 import 'package:felloapp/util/icici_api_util.dart';
 import 'package:felloapp/util/locator.dart';
 import 'package:felloapp/util/logger.dart';
-import 'package:felloapp/util/mixpanel_events.dart';
+import 'package:felloapp/core/service/analytics/analytics_events.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
@@ -43,7 +44,7 @@ class AugmontModel extends ChangeNotifier {
   final UserService _userService = locator<UserService>();
   final _userCoinService = locator<UserCoinService>();
   final TransactionService _txnService = locator<TransactionService>();
-  final MixpanelService _mixpanelService = locator<MixpanelService>();
+  final _analyticsService = locator<AnalyticsService>();
 
   ValueChanged<UserTransaction> _augmontTxnProcessListener;
   final String defaultBaseUri =
@@ -275,7 +276,7 @@ class AugmontModel extends ChangeNotifier {
         'BlockID: ${buyRates.blockId},gPrice: ${buyRates.goldBuyPrice}';
     String _note2 =
         'UserId:${_baseProvider.myUser.uid},MerchantTxnID: ${_tranIdResponse.model}';
-    String rzpOrderId = await _rzpGateway.createOrderId(amount, _note1, _note2);
+    String rzpOrderId = await _rzpGateway.createOrderId(amount, _baseProvider.myUser.uid, _tranIdResponse.model, _note1, _note2);
     if (rzpOrderId == null) {
       _logger.e("Received null from create Order id");
       return null;
@@ -374,8 +375,8 @@ class AugmontModel extends ChangeNotifier {
     if (_baseProvider.currentAugmontTxn.rzp[UserTransaction.subFldRzpStatus] ==
         UserTransaction.RZP_TRAN_STATUS_COMPLETE) {
       //payment completed successfully
-      _mixpanelService.track(
-          eventName: MixpanelEvents.investedInGold,
+      _analyticsService.track(
+          eventName: AnalyticsEvents.investedInGold,
           properties: {'goldQuantity': goldTxn.amount});
       _onPaymentComplete();
     } else {
@@ -466,6 +467,10 @@ class AugmontModel extends ChangeNotifier {
       _baseProvider.currentAugmontTxn = _onCompleteDepositResponse
           .model.response.transactionDoc.transactionDetail;
 
+      if (_onCompleteDepositResponse.model.gtId != null) {
+        GoldenTicketService.goldenTicketId =
+            _onCompleteDepositResponse.model.gtId;
+      }
       _txnService.updateTransactions();
 
       if (_augmontTxnProcessListener != null)
