@@ -1,24 +1,28 @@
-import 'package:felloapp/core/service/mixpanel_service.dart';
 import 'package:felloapp/main.dart';
 import 'package:felloapp/util/credentials_stage.dart';
 import 'package:felloapp/util/flavor_config.dart';
 import 'package:flutter/material.dart';
 
+import 'core/service/analytics/mixpanel_analytics.dart';
+
 void main() async {
   FlavorConfig(
-      flavor: Flavor.PROD,
-      color: Colors.deepPurpleAccent,
-      values: FlavorValues(
-          awsAugmontStage: AWSAugmontStage.PROD,
-          awsIciciStage: AWSIciciStage.PROD,
-          freshchatStage: FreshchatStage.DEV,
-          razorpayStage: RazorpayStage.PROD,
-          signzyStage: SignzyStage.PROD,
-          signzyPanStage: SignzyPanStage.PROD,
-          baseUriUS: 'us-central1-fello-d3a9c.cloudfunctions.net',
-          baseUriAsia: 'asia-south1-fello-d3a9c.cloudfunctions.net',
-          mixpanelToken: MixpanelService.PROD_TOKEN,
-          dynamicLinkPrefix: 'https://fello.in'));
+    flavor: Flavor.PROD,
+    color: Colors.deepPurpleAccent,
+    values: FlavorValues(
+      awsAugmontStage: AWSAugmontStage.PROD,
+      awsIciciStage: AWSIciciStage.PROD,
+      freshchatStage: FreshchatStage.DEV,
+      razorpayStage: RazorpayStage.PROD,
+      signzyStage: SignzyStage.PROD,
+      signzyPanStage: SignzyPanStage.PROD,
+      baseUriUS: 'us-central1-fello-d3a9c.cloudfunctions.net',
+      baseUriAsia: 'asia-south1-fello-d3a9c.cloudfunctions.net',
+      mixpanelToken: MixpanelAnalytics.PROD_TOKEN,
+      dynamicLinkPrefix: 'https://fello.in',
+    ),
+  );
+
   await mainInit();
   runApp(MyApp());
 }
