@@ -5,7 +5,8 @@ import 'ui_pages.dart';
 class FelloParser extends RouteInformationParser<PageConfiguration> {
   @override
   Future<PageConfiguration> parseRouteInformation(
-      RouteInformation routeInformation) async {
+    RouteInformation routeInformation,
+  ) async {
     final uri = Uri.parse(routeInformation.location);
 
     if (uri.pathSegments.isEmpty) {
