@@ -6,8 +6,9 @@ import 'package:flutter/material.dart';
 
 class BaseModel extends ChangeNotifier {
   ViewState _state = ViewState.Idle;
-
   ViewState get state => _state;
+
+  bool _disposed = false;
 
   void setState(ViewState viewState) {
     _state = viewState;
@@ -16,5 +17,18 @@ class BaseModel extends ChangeNotifier {
 
   void refresh() {
     notifyListeners();
+  }
+
+  @override
+  void notifyListeners() {
+    if (!_disposed) {
+      super.notifyListeners();
+    }
+  }
+
+  @override
+  void dispose() {
+    _disposed = true;
+    super.dispose();
   }
 }
