@@ -6,6 +6,8 @@ import 'package:felloapp/core/model/tambola_winners_details.dart';
 import 'package:felloapp/core/model/winners_model.dart';
 import 'package:felloapp/core/ops/lcl_db_ops.dart';
 import 'package:felloapp/core/repository/winners_repo.dart';
+import 'package:felloapp/core/service/analytics/analytics_events.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
 import 'package:felloapp/core/service/leaderboard_service.dart';
 import 'package:felloapp/core/service/user_service.dart';
 import 'package:felloapp/core/service/winners_service.dart';
@@ -17,6 +19,7 @@ import 'package:felloapp/util/locator.dart';
 import 'package:felloapp/util/styles/size_config.dart';
 import 'package:flutter/material.dart';
 import 'package:felloapp/util/custom_logger.dart';
+import 'package:sliding_up_panel/sliding_up_panel.dart';
 
 class WinViewModel extends BaseModel {
   final _userService = locator<UserService>();
@@ -24,10 +27,25 @@ class WinViewModel extends BaseModel {
   final _logger = locator<CustomLogger>();
   final _winnerService = locator<WinnerService>();
   final _lbService = locator<LeaderboardService>();
+  final _analyticsService = locator<AnalyticsService>();
 
   LocalDBModel _localDBModel = locator<LocalDBModel>();
   bool isWinnersLoading = false;
   WinnersModel _winners;
+  int _currentPage = 0;
+  int get getCurrentPage => this._currentPage;
+  static PanelController _panelController = PanelController();
+  PanelController get panelController => _panelController;
+
+  set panelController(val) {
+    _panelController = val;
+    notifyListeners();
+  }
+
+  set setCurrentPage(int currentPage) {
+    this._currentPage = currentPage;
+    notifyListeners();
+  }
 
   WinnersModel get winners => _winners;
 
@@ -43,7 +61,7 @@ class WinViewModel extends BaseModel {
 
   init() {}
 
-  getWinningsButtonText() {
+  String getWinningsButtonText() {
     if (_userService.userFundWallet.isPrizeBalanceUnclaimed())
       return "Redeem";
     else
@@ -59,6 +77,22 @@ class WinViewModel extends BaseModel {
         PageAction(state: PageState.addPage, page: MyWinnigsPageConfig);
   }
 
+  void navigateToRefer() {
+    _analyticsService.track(eventName: AnalyticsEvents.winReferral);
+    AppState.delegate.appState.currentAction = PageAction(
+      state: PageState.addPage,
+      page: ReferralDetailsPageConfig,
+    );
+  }
+
+  void navigateToWinnings() {
+    _analyticsService.track(eventName: AnalyticsEvents.winReferral);
+    AppState.delegate.appState.currentAction = PageAction(
+      state: PageState.addPage,
+      page: MyWinnigsPageConfig,
+    );
+  }
+
   openVoucherModal(
     String asset,
     String title,
@@ -71,20 +105,21 @@ class WinViewModel extends BaseModel {
       return;
     else
       return BaseUtil.openModalBottomSheet(
-          addToScreenStack: true,
-          content: VoucherModal(
-            color: color,
-            asset: asset,
-            commingSoon: commingsoon,
-            title: title,
-            subtitle: subtitle,
-            instructions: instructions,
-          ),
-          borderRadius: BorderRadius.only(
-              topLeft: Radius.circular(SizeConfig.padding24),
-              topRight: Radius.circular(SizeConfig.padding24)),
-          // backgroundColor: Color(0xffFFDBF6),
-          isBarrierDismissable: false,
-          hapticVibrate: true);
+        addToScreenStack: true,
+        content: VoucherModal(
+          color: color,
+          asset: asset,
+          commingSoon: commingsoon,
+          title: title,
+          subtitle: subtitle,
+          instructions: instructions,
+        ),
+        borderRadius: BorderRadius.only(
+            topLeft: Radius.circular(SizeConfig.padding24),
+            topRight: Radius.circular(SizeConfig.padding24)),
+        // backgroundColor: Color(0xffFFDBF6),
+        isBarrierDismissable: false,
+        hapticVibrate: true,
+      );
   }
 }
