@@ -10,10 +10,11 @@ import 'package:felloapp/core/model/base_user_model.dart';
 import 'package:felloapp/core/ops/augmont_ops.dart';
 import 'package:felloapp/core/ops/db_ops.dart';
 import 'package:felloapp/core/ops/lcl_db_ops.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
 import 'package:felloapp/core/service/api_service.dart';
 import 'package:felloapp/core/service/cache_manager.dart';
 import 'package:felloapp/core/service/fcm/fcm_listener_service.dart';
-import 'package:felloapp/core/service/mixpanel_service.dart';
+import 'package:felloapp/core/service/golden_ticket_service.dart';
 import 'package:felloapp/core/service/user_service.dart';
 import 'package:felloapp/navigator/app_state.dart';
 import 'package:felloapp/navigator/router/ui_pages.dart';
@@ -30,7 +31,7 @@ import 'package:felloapp/util/haptic.dart';
 import 'package:felloapp/util/localization/generated/l10n.dart';
 import 'package:felloapp/util/locator.dart';
 import 'package:felloapp/util/logger.dart';
-import 'package:felloapp/util/mixpanel_events.dart';
+import 'package:felloapp/core/service/analytics/analytics_events.dart';
 import 'package:felloapp/util/styles/size_config.dart';
 import 'package:felloapp/util/styles/textStyles.dart';
 import 'package:felloapp/util/styles/ui_constants.dart';
@@ -76,7 +77,7 @@ class _LoginControllerState extends State<LoginController>
   final UserService userService = locator<UserService>();
   final FcmListener fcmListener = locator<FcmListener>();
   final AugmontModel augmontProvider = locator<AugmontModel>();
-  final MixpanelService _mixpanelService = locator<MixpanelService>();
+  final _analyticsService = locator<AnalyticsService>();
   final _userService = locator<UserService>();
   final _logger = locator<CustomLogger>();
   final _apiPaths = locator<ApiPath>();
@@ -151,9 +152,11 @@ class _LoginControllerState extends State<LoginController>
       if (baseProvider.isOtpResendCount == 0) {
         ///this is the first time that the otp was requested
         baseProvider.isLoginNextInProgress = false;
-        _controller.animateToPage(OtpInputScreen.index,
-            duration: Duration(milliseconds: 500),
-            curve: Curves.easeInToLinear);
+        _controller.animateToPage(
+          OtpInputScreen.index,
+          duration: Duration(milliseconds: 500),
+          curve: Curves.easeInToLinear,
+        );
         setState(() {});
       } else {
         ///the otp was requested to be resent
@@ -211,12 +214,13 @@ class _LoginControllerState extends State<LoginController>
     };
 
     await FirebaseAuth.instance.verifyPhoneNumber(
-        phoneNumber: this._verificationId,
-        codeAutoRetrievalTimeout: autoRetrieve,
-        codeSent: smsCodeSent,
-        timeout: const Duration(seconds: 30),
-        verificationCompleted: verifiedSuccess,
-        verificationFailed: veriFailed);
+      phoneNumber: this._verificationId,
+      codeAutoRetrievalTimeout: autoRetrieve,
+      codeSent: smsCodeSent,
+      timeout: const Duration(seconds: 30),
+      verificationCompleted: verifiedSuccess,
+      verificationFailed: veriFailed,
+    );
   }
 
   @override
@@ -437,6 +441,10 @@ class _LoginControllerState extends State<LoginController>
                   'Only dummy numbers are allowed in QA mode');
               break;
             }
+
+            _analyticsService.track(
+              eventName: AnalyticsEvents.signupEnterMobile,
+            );
             this._verificationId = '+91' + this.userMobile;
             _verifyPhone();
             baseProvider.isLoginNextInProgress = true;
@@ -454,14 +462,14 @@ class _LoginControllerState extends State<LoginController>
             bool flag = await baseProvider.authenticateUser(baseProvider
                 .generateAuthCredential(_augmentedVerificationId, otp));
             if (flag) {
-              _mixpanelService.track(eventName: MixpanelEvents.mobileOtpDone);
+              _analyticsService.track(eventName: AnalyticsEvents.mobileOtpDone);
               AppState.isOnboardingInProgress = true;
               _otpScreenKey.currentState.onOtpReceived();
               _onSignInSuccess();
             } else {
               _otpScreenKey.currentState.pinEditingController.text = "";
               BaseUtil.showNegativeAlert(
-                  'Invalid Otp', 'Please enter a valid otp');
+                  'Incorrect OTP entered', 'Please enter a valid otp');
               baseProvider.isLoginNextInProgress = false;
               FocusScope.of(_otpScreenKey.currentContext).unfocus();
               setState(() {});
@@ -554,9 +562,10 @@ class _LoginControllerState extends State<LoginController>
               baseProvider.isLoginNextInProgress = false;
               setState(() {});
             }).then((value) {
-              _mixpanelService.track(
-                  eventName: MixpanelEvents.profileInformationAdded,
-                  properties: {'userId': baseProvider?.myUser?.uid});
+              _analyticsService.track(
+                eventName: AnalyticsEvents.profileInformationAdded,
+                properties: {'userId': baseProvider?.myUser?.uid},
+              );
               _controller.animateToPage(Username.index,
                   duration: Duration(milliseconds: 500),
                   curve: Curves.easeInToLinear);
@@ -609,6 +618,10 @@ class _LoginControllerState extends State<LoginController>
                         body: _body,
                         token: _bearer);
                     res['flag'] ? flag = true : flag = false;
+                    _logger.d("Is Golden Ticket Rewarded: ${res['gtId']}");
+                    if (res['gtId'] != null &&
+                        res['gtId'].toString().isNotEmpty)
+                      GoldenTicketService.goldenTicketId = res['gtId'];
                   } catch (e) {
                     _logger.d(e);
                     _usernameKey.currentState.enabled = true;
@@ -618,9 +631,11 @@ class _LoginControllerState extends State<LoginController>
                   // bool flag = await dbProvider.updateUser(baseProvider.myUser);
 
                   if (flag) {
-                    _mixpanelService.track(
-                        eventName: MixpanelEvents.userNameAdded,
-                        properties: {'userId': baseProvider?.myUser?.uid});
+                    _analyticsService.track(
+                      eventName: AnalyticsEvents.userNameAdded,
+                      properties: {'userId': baseProvider?.myUser?.uid},
+                    );
+
                     log.debug("User object saved successfully");
                     _onSignUpComplete();
                   } else {
@@ -755,16 +770,21 @@ class _LoginControllerState extends State<LoginController>
   }
 
   Future _onSignUpComplete() async {
-    if (_isSignup)
+    if (_isSignup) {
       await BaseAnalytics.analytics.logSignUp(signUpMethod: 'phonenumber');
+      _analyticsService.track(eventName: AnalyticsEvents.signupComplete);
+      _analyticsService.trackSignup(baseProvider.myUser.uid);
+    }
+
     await BaseAnalytics.logUserProfile(baseProvider.myUser);
     await userService.init();
     await baseProvider.init();
     await fcmListener.setupFcm();
-    _logger.i("Calling mixpanel init for new onborded user");
-    await _mixpanelService.init(
-        isOnboarded: userService.isUserOnborded,
-        baseUser: userService.baseUser);
+    _logger.i("Calling analytics login for new onborded user");
+    await _analyticsService.login(
+      isOnboarded: userService.isUserOnborded,
+      baseUser: userService.baseUser,
+    );
     AppState.isOnboardingInProgress = false;
     if (baseProvider.isLoginNextInProgress == true) {
       baseProvider.isLoginNextInProgress = false;
