@@ -153,6 +153,7 @@ class NameInputScreenState extends State<NameInputScreen> {
   }
 
   void _showAndroidDatePicker() async {
+    FocusScope.of(context).unfocus();
     var res = await showDatePicker(
       context: context,
       initialDate: DateTime(2000, 1, 1),
@@ -778,13 +779,13 @@ class _SignInOptionsState extends State<SignInOptions> {
                       color: UiConstants.primaryColor,
                     ),
                     title: Text("Use another email"),
-                    subtitle: Text(
-                      "this option requires an extra step",
-                      style: TextStyle(
-                        fontSize: SizeConfig.smallTextSize * 1.3,
-                        color: Colors.red[300],
-                      ),
-                    ),
+                    // subtitle: Text(
+                    //   "this option requires an extra step",
+                    //   style: TextStyle(
+                    //     fontSize: SizeConfig.smallTextSize * 1.3,
+                    //     color: Colors.red[300],
+                    //   ),
+                    // ),
                     onTap: () {
                       if (!baseProvider.isGoogleSignInProgress) {
                         widget.onEmailSignIn();
