@@ -49,6 +49,12 @@ class UsernameState extends State<Username> {
   UsernameResponse response;
 
   @override
+  void initState() {
+    focusNode = new FocusNode();
+    super.initState();
+  }
+
+  @override
   void dispose() {
     focusNode?.dispose();
     super.dispose();
@@ -138,15 +144,14 @@ class UsernameState extends State<Username> {
     );
   }
 
-  @override
-  void didChangeDependencies() {
-    if (mounted)
-      Future.delayed(Duration(seconds: 2), () {
-        if (mounted) FocusScope.of(context).requestFocus(focusNode);
-      });
+  // @override
+  // void didChangeDependencies() {
+  //   WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+  //     if (mounted) FocusScope.of(context).requestFocus();
+  //   });
 
-    super.didChangeDependencies();
-  }
+  //   super.didChangeDependencies();
+  // }
 
   @override
   Widget build(BuildContext context) {
