@@ -1,7 +1,12 @@
+import 'package:felloapp/core/service/analytics/analytics_events.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
+import 'package:felloapp/core/service/golden_ticket_service.dart';
 import 'package:felloapp/navigator/app_state.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_scratch_dialog/gt_instant_view.dart';
 import 'package:felloapp/ui/pages/static/home_background.dart';
 import 'package:felloapp/ui/widgets/buttons/fello_button/large_button.dart';
 import 'package:felloapp/util/assets.dart';
+import 'package:felloapp/util/locator.dart';
 import 'package:felloapp/util/styles/size_config.dart';
 import 'package:felloapp/util/styles/textStyles.dart';
 import 'package:felloapp/util/styles/ui_constants.dart';
@@ -17,6 +22,8 @@ class _WalkThroughPageState extends State<WalkThroughPage> {
   PageController _pageController;
   ValueNotifier<double> _pageNotifier;
   bool showLotties = false;
+  GoldenTicketService _gtService = GoldenTicketService();
+  final _analyticsService = locator<AnalyticsService>();
 
   List<String> lottieList = [Assets.onb1, Assets.onb2, Assets.onb3];
   List<String> titleList = ["SAVE", "PLAY", "WIN"];
@@ -25,6 +32,7 @@ class _WalkThroughPageState extends State<WalkThroughPage> {
     "Use these tokens to play fun and exciting games 🎮",
     "Stand to win exclusive prizes and fun rewards 🎉"
   ];
+
   @override
   void initState() {
     _pageController = PageController();
@@ -35,6 +43,7 @@ class _WalkThroughPageState extends State<WalkThroughPage> {
         showLotties = true;
       });
     });
+    _analyticsService.track(eventName: AnalyticsEvents.signupDemo);
     super.initState();
   }
 
@@ -216,11 +225,21 @@ class _WalkThroughPageState extends State<WalkThroughPage> {
                         style: TextStyles.body2.bold.colour(Colors.white),
                       ),
                       onPressed: () {
-                        value.toInt() == 2
-                            ? AppState.backButtonDispatcher.didPopRoute()
-                            : _pageController.nextPage(
-                                duration: Duration(milliseconds: 400),
-                                curve: Curves.easeIn);
+                        if (value.toInt() == 2) {
+                          AppState.backButtonDispatcher.didPopRoute();
+                          _gtService
+                              .fetchAndVerifyGoldenTicketByID()
+                              .then((bool res) {
+                            if (res)
+                              _gtService.showInstantGoldenTicketView(
+                                  title: 'Welcome to Fello',
+                                  source: GTSOURCE.newuser);
+                          });
+                        } else
+                          _pageController.nextPage(
+                            duration: Duration(milliseconds: 400),
+                            curve: Curves.easeIn,
+                          );
                       },
                     );
                   },
@@ -229,6 +248,12 @@ class _WalkThroughPageState extends State<WalkThroughPage> {
               TextButton(
                 onPressed: () {
                   AppState.backButtonDispatcher.didPopRoute();
+                  _gtService.fetchAndVerifyGoldenTicketByID().then((bool res) {
+                    if (res)
+                      _gtService.showInstantGoldenTicketView(
+                          title: 'Welcome to Fello',
+                          source: GTSOURCE.newuser);
+                  });
                 },
                 child: Text(
                   "Skip",
@@ -237,7 +262,7 @@ class _WalkThroughPageState extends State<WalkThroughPage> {
               ),
               SizedBox(
                 height: kToolbarHeight / 2,
-              )
+              ),
             ],
           ),
         ),
