@@ -2,6 +2,8 @@ import 'package:felloapp/base_util.dart';
 import 'package:felloapp/core/model/leader_board_modal.dart';
 import 'package:felloapp/core/model/prizes_model.dart';
 import 'package:felloapp/core/repository/statistics_repo.dart';
+import 'package:felloapp/core/service/analytics/analytics_events.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
 import 'package:felloapp/core/service/prize_service.dart';
 import 'package:felloapp/ui/architecture/base_vm.dart';
 import 'package:felloapp/util/api_response.dart';
@@ -13,6 +15,7 @@ class TambolaHomeViewModel extends BaseModel {
   final _stats = locator<StatisticsRepository>();
   final _prizeService = locator<PrizeService>();
   final _logger = locator<CustomLogger>();
+  final _analyticsService = locator<AnalyticsService>();
 
   bool isLeaderboardLoading = false;
   bool isPrizesLoading = false;
@@ -78,7 +81,8 @@ class TambolaHomeViewModel extends BaseModel {
     notifyListeners();
   }
 
-  openGame() {
+  void openGame() {
+    _analyticsService.track(eventName: AnalyticsEvents.playsTambola);
     BaseUtil().openTambolaHome();
   }
 }
