@@ -1,8 +1,12 @@
+import 'dart:ui';
 
+import 'package:felloapp/core/enums/page_state_enum.dart';
+import 'package:felloapp/core/service/golden_ticket_service.dart';
+import 'package:felloapp/navigator/app_state.dart';
+import 'package:felloapp/navigator/router/ui_pages.dart';
 import 'package:felloapp/ui/architecture/base_view.dart';
-import 'package:felloapp/ui/pages/others/games/cricket/cricket_home/cricket_home_view.dart';
-import 'package:felloapp/ui/pages/others/games/tambola/tambola_home/tambola_home_view.dart';
 import 'package:felloapp/ui/pages/others/profile/my_winnings/my_winnings_vm.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_tickets/golden_tickets_view.dart';
 import 'package:felloapp/ui/pages/static/fello_appbar.dart';
 import 'package:felloapp/ui/pages/static/home_background.dart';
 import 'package:felloapp/ui/pages/static/winnings_container.dart';
@@ -13,8 +17,8 @@ import 'package:felloapp/util/styles/size_config.dart';
 import 'package:felloapp/util/styles/textStyles.dart';
 import 'package:felloapp/util/styles/ui_constants.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:intl/intl.dart';
 
 class MyWinningsView extends StatelessWidget {
   @override
@@ -23,9 +27,11 @@ class MyWinningsView extends StatelessWidget {
     return BaseView<MyWinningsViewModel>(
       onModelReady: (model) {
         model.getWinningHistory();
+        // model.getGoldenTickets();
       },
       builder: (ctx, model, child) {
         return Scaffold(
+          //floatingActionButton: AddTodoButton(),
           backgroundColor: UiConstants.primaryColor,
           body: HomeBackground(
             child: Column(
@@ -41,274 +47,87 @@ class MyWinningsView extends StatelessWidget {
                       topRight: Radius.circular(SizeConfig.padding40),
                     ),
                     child: Container(
-                      decoration: BoxDecoration(
-                        color: Colors.white,
-                      ),
-                      padding: EdgeInsets.only(
-                          top: SizeConfig.pageHorizontalMargins),
-                      child: ListView(
-                        padding: EdgeInsets.zero,
-                        //crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Hero(
-                            tag: "myWinnings",
-                            child: WinningsContainer(
-                              shadow: false,
-                            ),
-                          ),
-                          SizedBox(height: SizeConfig.padding24),
-                          PrizeClaimCard(
-                            model: model,
-                          ),
-                          Padding(
-                            padding: EdgeInsets.symmetric(
-                                horizontal: SizeConfig.pageHorizontalMargins),
-                            child: Text(
-                              "Winning History",
-                              style: TextStyles.title3.bold,
-                            ),
-                          ),
-                          SizedBox(height: SizeConfig.padding16),
-                          model.isWinningHistoryLoading
-                              ? ListLoader()
-                              : (model.winningHistory != null &&
-                                      model.winningHistory.isNotEmpty
-                                  ? Container(
-                                      padding: EdgeInsets.symmetric(
-                                          horizontal:
-                                              SizeConfig.pageHorizontalMargins /
-                                                  4),
-                                      child: Column(
-                                        children: List.generate(
-                                          model.winningHistory.length,
-                                          (i) => Theme(
-                                            data: ThemeData().copyWith(
-                                                dividerColor: Colors.grey[50]),
-                                            child: ExpansionTile(
-                                              expandedCrossAxisAlignment:
-                                                  CrossAxisAlignment.start,
-                                              expandedAlignment:
-                                                  Alignment.centerLeft,
-                                              children: [
-                                                Container(
-                                                  margin: EdgeInsets.only(
-                                                      bottom:
-                                                          SizeConfig.padding8),
-                                                  padding: EdgeInsets.only(
-                                                    left: SizeConfig
-                                                            .pageHorizontalMargins +
-                                                        SizeConfig.padding20 *
-                                                            2 +
-                                                        SizeConfig.padding8,
-                                                    right: SizeConfig
-                                                            .pageHorizontalMargins /
-                                                        2,
-                                                  ),
-                                                  child: Row(
-                                                    children: [
-                                                      Column(
-                                                        crossAxisAlignment:
-                                                            CrossAxisAlignment
-                                                                .start,
-                                                        children: [
-                                                          Row(
-                                                            children: [
-                                                              Text("Type: "),
-                                                              Text((model.winningHistory[i].redeemType !=
-                                                                          null &&
-                                                                      model.winningHistory[i]
-                                                                              .redeemType !=
-                                                                          "")
-                                                                  ? "REDEMPTION"
-                                                                  : "CREDIT"),
-                                                            ],
-                                                          ),
-                                                          SizedBox(
-                                                              height: SizeConfig
-                                                                  .padding4),
-                                                          Row(
-                                                            children: [
-                                                              Text("Status: "),
-                                                              Text(model
-                                                                      .winningHistory[
-                                                                          i]
-                                                                      .tranStatus ??
-                                                                  "COMPLETED"),
-                                                            ],
-                                                          )
-                                                        ],
-                                                      ),
-                                                      Spacer(),
-                                                      if (model
-                                                                  .winningHistory[
-                                                                      i]
-                                                                  .redeemType !=
-                                                              null &&
-                                                          model.winningHistory[i]
-                                                                  .redeemType !=
-                                                              "")
-                                                        InkWell(
-                                                          onTap: () => model.showPrizeDetailsDialog(
-                                                              model
-                                                                      .winningHistory[
-                                                                          i]
-                                                                      .redeemType ??
-                                                                  "",
-                                                              model.winningHistory[i]
-                                                                      .amount
-                                                                      .abs() ??
-                                                                  0.0),
-                                                          child: Container(
-                                                            padding: EdgeInsets.symmetric(
-                                                                horizontal:
-                                                                    SizeConfig
-                                                                        .padding12,
-                                                                vertical:
-                                                                    SizeConfig
-                                                                        .padding8),
-                                                            decoration:
-                                                                BoxDecoration(
-                                                              borderRadius:
-                                                                  BorderRadius
-                                                                      .circular(
-                                                                          8),
-                                                              color: UiConstants
-                                                                  .primaryColor,
-                                                            ),
-                                                            child: Row(
-                                                              mainAxisAlignment:
-                                                                  MainAxisAlignment
-                                                                      .center,
-                                                              children: [
-                                                                SvgPicture
-                                                                    .asset(
-                                                                  Assets.plane,
-                                                                  color: Colors
-                                                                      .white,
-                                                                  width: SizeConfig
-                                                                      .padding12,
-                                                                ),
-                                                                SizedBox(
-                                                                    width: SizeConfig
-                                                                        .padding8),
-                                                                Text(
-                                                                  "Share",
-                                                                  style: TextStyles
-                                                                      .body3
-                                                                      .colour(Colors
-                                                                          .white),
-                                                                )
-                                                              ],
-                                                            ),
-                                                          ),
-                                                        ),
-                                                    ],
-                                                  ),
-                                                )
-                                              ],
-                                              textColor: Colors.black,
-                                              leading: CircleAvatar(
-                                                radius: SizeConfig.padding24,
-                                                backgroundColor: model
-                                                    .getWinningHistoryLeadingBg(
-                                                        model.winningHistory[i]
-                                                                .redeemType ??
-                                                            ""),
-                                                child: Padding(
-                                                  padding: EdgeInsets.all(
-                                                      SizeConfig.padding12),
-                                                  child: Container(
-                                                    decoration: BoxDecoration(
-                                                        shape: BoxShape.circle,
-                                                        boxShadow: [
-                                                          BoxShadow(
-                                                            color: Colors.black
-                                                                .withOpacity(
-                                                                    0.2),
-                                                            blurRadius: 2,
-                                                            offset:
-                                                                Offset(4, 4),
-                                                            spreadRadius: 2,
-                                                          )
-                                                        ]),
-                                                    child: Image.asset(model
-                                                        .getWinningHistoryLeadingImage(
-                                                            model.winningHistory[i]
-                                                                    .redeemType ??
-                                                                "")),
-                                                  ),
-                                                ),
-                                              ),
-                                              title: Text(
-                                                model.getWinningHistoryTitle(
-                                                    model.winningHistory[i]),
-                                                style: TextStyles.body2.bold,
-                                              ),
-                                              subtitle: Row(
-                                                children: [
-                                                  Text(
-                                                    DateFormat("dd MMM, yyyy")
-                                                        .format(model
-                                                            .winningHistory[i]
-                                                            .timestamp
-                                                            .toDate()),
-                                                    style: TextStyles.body3
-                                                        .colour(Colors.grey),
-                                                  ),
-                                                  SizedBox(width: 10),
-                                                  if (model.winningHistory[i]
-                                                          .tranStatus ==
-                                                      "PROCESSING")
-                                                    Container(
-                                                      decoration: BoxDecoration(
-                                                        borderRadius:
-                                                            BorderRadius
-                                                                .circular(100),
-                                                        color: Colors.yellow,
-                                                      ),
-                                                      padding:
-                                                          EdgeInsets.symmetric(
-                                                              horizontal: 8,
-                                                              vertical: 2),
-                                                      child: Text(
-                                                        model.winningHistory[i]
-                                                            .tranStatus,
-                                                        style: TextStyles
-                                                            .body4.bold
-                                                            .colour(
-                                                                Colors.white),
-                                                      ),
-                                                    )
-                                                ],
-                                              ),
-                                              trailing: Text(
-                                                model.txnService
-                                                    .getFormattedTxnAmount(model
-                                                        .winningHistory[i]
-                                                        .amount),
-                                                style: TextStyles.body2.bold
-                                                    .colour(
-                                                        model.winningHistory[i]
-                                                                    .amount >
-                                                                0
-                                                            ? UiConstants
-                                                                .primaryColor
-                                                            : Colors.blue[700]),
-                                              ),
-                                            ),
-                                          ),
+                        decoration: BoxDecoration(
+                          color: Colors.white,
+                        ),
+                        child: NestedScrollView(
+                          // allows you to build a list of elements that would be scrolled away till the body reached the top
+                          headerSliverBuilder: (context, _) {
+                            return [
+                              SliverList(
+                                delegate: SliverChildListDelegate([
+                                  PrizeClaimCard(
+                                    model: model,
+                                  ),
+                                ]),
+                              ),
+                            ];
+                          },
+                          body: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              SizedBox(height: SizeConfig.padding24),
+                              WinningsContainer(
+                                shadow: false,
+                                onTap: () {
+                                  AppState.delegate.appState.currentAction =
+                                      PageAction(
+                                          state: PageState.addPage,
+                                          page: GoldenMilestonesViewPageConfig);
+                                },
+                                child: Container(
+                                  child: Padding(
+                                    padding:
+                                        EdgeInsets.all(SizeConfig.padding16),
+                                    child: Row(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.center,
+                                      children: [
+                                        SvgPicture.asset(
+                                          Assets.giftBoxOpen,
                                         ),
-                                      ),
-                                    )
-                                  : Center(
-                                      child: NoRecordDisplayWidget(
-                                        asset: Assets.noTransaction,
-                                        text: "No Winning History yet",
-                                      ),
-                                    ))
-                        ],
-                      ),
-                    ),
+                                        SizedBox(
+                                            width:
+                                                SizeConfig.screenWidth * 0.05),
+                                        Column(
+                                          crossAxisAlignment:
+                                              CrossAxisAlignment.start,
+                                          mainAxisAlignment:
+                                              MainAxisAlignment.spaceEvenly,
+                                          children: [
+                                            Text(
+                                              "Earn your next\n Golden Tickets",
+                                              style: TextStyles.body1
+                                                  .colour(Colors.white)
+                                                  .light,
+                                            ),
+                                          ],
+                                        )
+                                      ],
+                                    ),
+                                  ),
+                                ),
+                              ),
+                              Padding(
+                                padding: EdgeInsets.only(
+                                    top: SizeConfig.padding16,
+                                    left: SizeConfig.pageHorizontalMargins),
+                                child: Text(
+                                  "My Rewards",
+                                  style: TextStyles.title3.bold,
+                                ),
+                              ),
+                              GoldenTicketsView()
+                            ],
+                          ),
+                        )
+                        //  ListView(
+                        //   padding: EdgeInsets.zero,
+                        //   // crossAxisAlignment: CrossAxisAlignment.start,
+                        //   children: [
+
+                        // ),
+                        ),
                   ),
                 )
               ],
