@@ -2,14 +2,14 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:felloapp/base_util.dart';
 import 'package:felloapp/core/model/referral_details_model.dart';
 import 'package:felloapp/core/ops/db_ops.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
 import 'package:felloapp/ui/pages/others/games/cricket/cricket_home/cricket_home_view.dart';
-import 'package:felloapp/core/service/mixpanel_service.dart';
 import 'package:felloapp/ui/pages/static/fello_appbar.dart';
 import 'package:felloapp/ui/pages/static/home_background.dart';
 import 'package:felloapp/util/assets.dart';
 import 'package:felloapp/util/constants.dart';
 import 'package:felloapp/util/locator.dart';
-import 'package:felloapp/util/mixpanel_events.dart';
+import 'package:felloapp/core/service/analytics/analytics_events.dart';
 import 'package:felloapp/util/styles/size_config.dart';
 import 'package:felloapp/util/styles/textStyles.dart';
 import 'package:felloapp/util/styles/ui_constants.dart';
@@ -29,7 +29,7 @@ class ReferralHistoryView extends StatefulWidget {
 class _ReferralHistoryViewState extends State<ReferralHistoryView> {
   BaseUtil baseProvider;
   DBModel dbProvider;
-  final _mixpanelService = locator<MixpanelService>();
+  final _analyticsService = locator<AnalyticsService>();
 
   @override
   Widget build(BuildContext context) {
@@ -55,8 +55,8 @@ class _ReferralHistoryViewState extends State<ReferralHistoryView> {
           if (baseProvider.myReferralInfo != null && _t < _n) {
             baseProvider.myReferralInfo.refCount = _n;
             if (_n != null && _n > 0)
-              _mixpanelService.track(
-                  eventName: MixpanelEvents.referralCount,
+              _analyticsService.track(
+                  eventName: AnalyticsEvents.referralCount,
                   properties: {"count": _n});
             dbProvider.updateUserReferralCount(baseProvider.myUser.uid,
                 baseProvider.myReferralInfo); //await not required
@@ -202,28 +202,28 @@ class _ReferralHistoryViewState extends State<ReferralHistoryView> {
                       ? UiConstants.primaryColor
                       : UiConstants.tertiarySolid),
                 ),
-                SizedBox(height: SizeConfig.padding2),
-                InkWell(
-                  onTap: () {
-                    BaseUtil.showPositiveAlert(
-                        "${rDetail.userName} pinged 👍🏼",
-                        "You can ping again after 48 hours");
-                  },
-                  child: Container(
-                    decoration: BoxDecoration(
-                      color: UiConstants.tertiarySolid,
-                      borderRadius:
-                          BorderRadius.circular(SizeConfig.roundness12),
-                    ),
-                    padding: EdgeInsets.symmetric(
-                        vertical: SizeConfig.padding6,
-                        horizontal: SizeConfig.padding12),
-                    child: Text(
-                      "Ping",
-                      style: TextStyles.body3.bold.colour(Colors.white),
-                    ),
-                  ),
-                )
+                // SizedBox(height: SizeConfig.padding2),
+                // InkWell(
+                //   onTap: () {
+                //     BaseUtil.showPositiveAlert(
+                //         "${rDetail.userName} pinged 👍🏼",
+                //         "You can ping again after 48 hours");
+                //   },
+                //   child: Container(
+                //     decoration: BoxDecoration(
+                //       color: UiConstants.tertiarySolid,
+                //       borderRadius:
+                //           BorderRadius.circular(SizeConfig.roundness12),
+                //     ),
+                //     padding: EdgeInsets.symmetric(
+                //         vertical: SizeConfig.padding6,
+                //         horizontal: SizeConfig.padding12),
+                //     child: Text(
+                //       "Ping",
+                //       style: TextStyles.body3.bold.colour(Colors.white),
+                //     ),
+                //   ),
+                // )
               ],
             ),
           ],
@@ -235,19 +235,18 @@ class _ReferralHistoryViewState extends State<ReferralHistoryView> {
         rDetail.isRefereeBonusUnlocked ||
         rDetail.isUserBonusUnlocked == null ||
         rDetail.isUserBonusUnlocked);
-    if (!_isBonusUnlocked)
-      return 'Not yet invested 🔒';
-    else {
-      if (rDetail.bonusMap != null &&
-          rDetail.bonusMap['uamt'] != null &&
-          rDetail.bonusMap['uflc'] != null) {
-        int _amt = BaseUtil.toInt(rDetail.bonusMap['uamt']);
-        int _tck = BaseUtil.toInt(rDetail.bonusMap['uflc']);
-        if (_amt != null && _tck != null)
-          return 'You earned ₹$_amt and $_tck tokens 🥳';
-      }
-    }
-    return 'Rewards unlocked 🥳';
+    if (!_isBonusUnlocked) return 'Not yet invested 🔒';
+    // else {
+    //   if (rDetail.bonusMap != null &&
+    //       rDetail.bonusMap['uamt'] != null &&
+    //       rDetail.bonusMap['uflc'] != null) {
+    //     int _amt = BaseUtil.toInt(rDetail.bonusMap['uamt']);
+    //     int _tck = BaseUtil.toInt(rDetail.bonusMap['uflc']);
+    //     if (_amt != null && _tck != null)
+    //       return 'You earned ₹$_amt and $_tck tokens 🥳';
+    //   }
+    // }
+    return 'You earned a Golden Ticket 🥳';
   }
 
   String _getUserMembershipDate(Timestamp tmp) {
