@@ -7,6 +7,7 @@ import 'package:felloapp/core/ops/https/http_ops.dart';
 import 'package:felloapp/core/service/user_service.dart';
 import 'package:felloapp/navigator/app_state.dart';
 import 'package:felloapp/ui/elements/pin_input_custom_text_field.dart';
+import 'package:felloapp/ui/pages/login/screens/name_input_screen.dart';
 import 'package:felloapp/ui/pages/static/fello_appbar.dart';
 import 'package:felloapp/ui/pages/static/home_background.dart';
 import 'package:felloapp/util/locator.dart';
@@ -32,7 +33,7 @@ class VerifyEmailState extends State<VerifyEmail> {
   TextEditingController email = new TextEditingController();
   TextEditingController otp = new TextEditingController();
   final _userService = locator<UserService>();
-  final _baseUtil = locator<BaseUtil>();
+  // final baseProvider = locator<BaseUtil>();
   final formKey = GlobalKey<FormState>();
   Timer timer;
   bool isGmailVerifying = false;
@@ -41,7 +42,8 @@ class VerifyEmailState extends State<VerifyEmail> {
   String generatedOTP;
   HttpModel httpProvider;
   bool _isContinueWithGoogle = false;
-  //bool _baseUtil.isGoogleSignInProgress = false;
+  //bool baseProvider.isGoogleSignInProgress = false;
+  FocusNode focusNode;
   bool _isOtpSent = false;
   bool _isProcessing = false;
   bool _isVerifying = false;
@@ -51,8 +53,10 @@ class VerifyEmailState extends State<VerifyEmail> {
 
   @override
   void initState() {
-    email = TextEditingController(text: "1000");
+    email = TextEditingController(text: "");
     WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+      baseProvider.isGoogleSignInProgress = false;
+      focusNode = new FocusNode();
       showEmailOptions();
     });
     super.initState();
@@ -66,86 +70,24 @@ class VerifyEmailState extends State<VerifyEmail> {
   }
 
   showEmailOptions() {
-    showModalBottomSheet(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
-        isDismissible: false,
-        context: context,
-        builder: (ctx) {
-          return Wrap(
-            children: [
-              Container(
-                decoration: BoxDecoration(),
-                padding: EdgeInsets.all(
-                  SizeConfig.blockSizeHorizontal * 5,
-                ),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      children: [
-                        Text(
-                          "Verification option",
-                          style: TextStyle(
-                            color: Colors.black54,
-                            fontWeight: FontWeight.w700,
-                            fontSize: 24,
-                          ),
-                        ),
-                        Spacer(),
-                        IconButton(
-                          icon: Icon(Icons.close),
-                          iconSize: 30,
-                          onPressed: () =>
-                              AppState.backButtonDispatcher.didPopRoute(),
-                        ),
-                      ],
-                    ),
-                    Divider(
-                      height: 32,
-                      thickness: 2,
-                    ),
-                    ListTile(
-                      leading: SvgPicture.asset(
-                        "images/svgs/google.svg",
-                        height: 24,
-                        width: 24,
-                      ),
-                      title: Text("Choose a Google account"),
-                      subtitle: Text(""),
-                      onTap: verifyGmail,
-                      trailing: _baseUtil.isGoogleSignInProgress
-                          ? Container(
-                              child: CircularProgressIndicator(),
-                            )
-                          : SizedBox(),
-                    ),
-                    Divider(),
-                    ListTile(
-                      leading: Icon(
-                        Icons.alternate_email,
-                        color: UiConstants.primaryColor,
-                      ),
-                      title: Text("Receive an OTP on your email"),
-                      onTap: () {
-                        setState(() {
-                          _baseUtil.isGoogleSignInProgress = false;
-                          _isContinueWithGoogle = false;
-                          email.text = baseProvider.myUser.email;
-                          _isEmailEnabled = true;
-                        });
-                        Navigator.pop(context);
-                      },
-                    ),
-                    SizedBox(
-                      height: 24,
-                    )
-                  ],
-                ),
-                width: double.infinity,
-              ),
-            ],
-          );
-        });
+    baseProvider.isGoogleSignInProgress = false;
+    BaseUtil.openModalBottomSheet(
+        isBarrierDismissable: false,
+        borderRadius: BorderRadius.circular(15),
+        addToScreenStack: false,
+        hapticVibrate: true,
+        content: SignInOptions(
+          onGoogleSignIn: verifyGmail,
+          onEmailSignIn: () {
+            baseProvider.isGoogleSignInProgress = false;
+            // _isContinueWithGoogle = false;
+            email.text = baseProvider.myUser.email;
+            // _isEmailEnabled = true;
+
+            Navigator.pop(context);
+            focusNode.requestFocus();
+          },
+        ));
   }
 
   generateOtp() {
@@ -230,9 +172,8 @@ class VerifyEmailState extends State<VerifyEmail> {
   }
 
   verifyGmail() async {
-    setState(() {
-      _baseUtil.isGoogleSignInProgress = true;
-    });
+    baseProvider.isGoogleSignInProgress = true;
+
     final _gSignIn = GoogleSignIn();
     try {
       if (await _gSignIn.isSignedIn()) await _gSignIn.signOut();
@@ -251,27 +192,27 @@ class VerifyEmailState extends State<VerifyEmail> {
         bool res = await dbProvider.updateUser(baseProvider.myUser);
         if (res) {
           setState(() {
-            _baseUtil.isGoogleSignInProgress = false;
+            baseProvider.isGoogleSignInProgress = false;
           });
           BaseUtil.showPositiveAlert("Success", "Email Verified successfully");
           Navigator.pop(context);
           AppState.backButtonDispatcher.didPopRoute();
         } else {
-          _baseUtil.isGoogleSignInProgress = false;
+          baseProvider.isGoogleSignInProgress = false;
           BaseUtil.showNegativeAlert(
             "Oops! we ran into problem",
             "Email cannot be verified at the moment",
           );
         }
       } else {
-        _baseUtil.isGoogleSignInProgress = false;
+        baseProvider.isGoogleSignInProgress = false;
         BaseUtil.showNegativeAlert(
           "Email already registered",
           "Please try with another email",
         );
       }
     } else {
-      _baseUtil.isGoogleSignInProgress = false;
+      baseProvider.isGoogleSignInProgress = false;
       BaseUtil.showNegativeAlert(
         "No account selected",
         "Please choose an account from the list",
@@ -291,10 +232,11 @@ class VerifyEmailState extends State<VerifyEmail> {
 
   @override
   Widget build(BuildContext context) {
-    baseProvider = Provider.of<BaseUtil>(context, listen: false);
+    baseProvider = Provider.of<BaseUtil>(context);
     dbProvider = Provider.of<DBModel>(context, listen: false);
     httpProvider = Provider.of<HttpModel>(context, listen: false);
     return Scaffold(
+      resizeToAvoidBottomInset: false,
       body: HomeBackground(
         child: Column(
           children: [
@@ -315,217 +257,197 @@ class VerifyEmailState extends State<VerifyEmail> {
                 ),
                 child: Stack(
                   children: [
-                    Container(
-                      padding: EdgeInsets.fromLTRB(
-                          0,
-                          0,
-                          0,
-                          math.min(MediaQuery.of(context).viewInsets.bottom,
-                              SizeConfig.screenHeight * 0.09)),
-                      width: SizeConfig.screenWidth,
-                      // height: SizeConfig.screenHeight,
-                      child: SingleChildScrollView(
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          //crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            SizedBox(
-                              height: kToolbarHeight,
-                            ),
-                            Text(
-                              "Complete verification",
-                              style: TextStyle(
-                                  fontWeight: FontWeight.w700,
-                                  fontSize: SizeConfig.cardTitleTextSize),
-                            ),
-                            SizedBox(
-                              height: 24,
-                            ),
-                            Text(
-                                "Please enter the email where you'd like to receive all transaction related updates"),
-                            Form(
-                              key: formKey,
-                              child: Container(
-                                padding: EdgeInsets.only(top: 30, bottom: 10),
-                                child: TextFormField(
-                                  controller: email,
-                                  enabled: _isProcessing || _isOtpSent
-                                      ? false
-                                      : true,
-                                  cursorColor: Colors.black,
-                                  keyboardType: TextInputType.text,
-                                  validator: (val) {
-                                    if (val == "")
-                                      return null;
-                                    else if (val == null)
-                                      return "Please enter an email";
-                                    else if (!RegExp(
-                                            r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$')
-                                        .hasMatch(val)) {
-                                      return "Enter a valid email";
-                                    }
-                                    return null;
-                                  },
-                                  decoration: InputDecoration(
-                                    prefixIcon: Icon(Icons.email_rounded),
-                                  ),
-                                ),
+                    ListView(
+                      //mainAxisSize: MainAxisSize.min,
+                      //crossAxisAlignment: CrossAxisAlignment.start,
+                      padding: EdgeInsets.symmetric(
+                          vertical: SizeConfig.pageHorizontalMargins),
+                      children: [
+                        Text(
+                          "Email verification",
+                          style: TextStyle(
+                              fontWeight: FontWeight.w700,
+                              fontSize: SizeConfig.cardTitleTextSize),
+                        ),
+                        SizedBox(
+                          height: 24,
+                        ),
+                        Text(
+                            "Please enter the email where you would like to receive all transaction and support related updates"),
+                        Form(
+                          key: formKey,
+                          child: Container(
+                            padding: EdgeInsets.only(top: 30, bottom: 10),
+                            child: TextFormField(
+                              focusNode: focusNode,
+                              controller: email,
+                              enabled:
+                                  _isProcessing || _isOtpSent ? false : true,
+                              cursorColor: Colors.black,
+                              keyboardType: TextInputType.text,
+                              validator: (val) {
+                                if (val == "")
+                                  return null;
+                                else if (val == null)
+                                  return "Please enter an email";
+                                else if (!RegExp(
+                                        r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$')
+                                    .hasMatch(val)) {
+                                  return "Enter a valid email";
+                                }
+                                return null;
+                              },
+                              decoration: InputDecoration(
+                                prefixIcon: Icon(Icons.email_rounded),
                               ),
                             ),
-                            Padding(
-                              padding: EdgeInsets.symmetric(vertical: 24),
-                              child: Text((!_isOtpSent)
-                                  ? "We will send you a 6 digit code on this email."
-                                  : "Please check your inbox folders for the code"),
-                            ),
-                            SizedBox(
-                              height: 24,
-                            ),
-                            _isProcessing
-                                ? Container(
-                                    width: SizeConfig.screenWidth,
-                                    alignment: Alignment.center,
-                                    child: Column(
-                                      children: [
-                                        CircularProgressIndicator(),
-                                        SizedBox(
-                                          height: 8,
-                                        ),
-                                        Text("Sending OTP")
-                                      ],
-                                    ),
-                                  )
-                                : SizedBox(),
-                            _isOtpSent
-                                ? Container(
-                                    child: Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
-                                        Text(
-                                          "Enter the OTP",
-                                          style: TextStyle(
-                                              fontWeight: FontWeight.w700,
-                                              fontSize:
-                                                  SizeConfig.cardTitleTextSize),
-                                        ),
-                                        SizedBox(
-                                          height: 24,
-                                        ),
-                                        Padding(
-                                          padding: const EdgeInsets.fromLTRB(
-                                              0, 18.0, 0, 18.0),
-                                          child: PinInputTextField(
-                                            autoFocus: true,
-                                            pinLength: 6,
-                                            decoration: BoxLooseDecoration(
-                                              enteredColor:
-                                                  UiConstants.primaryColor,
-                                              solidColor: UiConstants
-                                                  .primaryColor
-                                                  .withOpacity(0.04),
-                                              strokeColor:
-                                                  UiConstants.primaryColor,
-                                              strokeWidth: 1,
-                                              textStyle: GoogleFonts.montserrat(
-                                                  fontSize: 20,
-                                                  color: Colors.black),
-                                            ),
-                                            controller: otp,
-                                            onChanged: (value) {
-                                              print(value);
-                                              if (value.length == 6) {
-                                                verifyOtp();
-                                              } else {
-                                                setState(() {
-                                                  _isOtpIncorrect = false;
-                                                });
-                                              }
-                                            },
-                                            onSubmit: (pin) {
-                                              print("Pressed submit for pin: " +
-                                                  pin.toString() +
-                                                  "\n  No action taken.");
-                                            },
-                                          ),
-                                        ),
-                                        SizedBox(
-                                          height: 24,
-                                        ),
-                                        Row(
-                                          children: [
-                                            Text("OTP is only valid for "),
-                                            TweenAnimationBuilder<Duration>(
-                                                duration: Duration(minutes: 15),
-                                                tween: Tween(
-                                                    begin:
-                                                        Duration(minutes: 15),
-                                                    end: Duration.zero),
-                                                onEnd: () {
-                                                  print('Timer ended');
-                                                  BaseUtil.showNegativeAlert(
-                                                    "Session Expired!",
-                                                    "Please try again",
-                                                  );
-                                                  AppState.backButtonDispatcher
-                                                      .didPopRoute();
-                                                },
-                                                builder: (BuildContext context,
-                                                    Duration value,
-                                                    Widget child) {
-                                                  final minutes =
-                                                      value.inMinutes;
-                                                  final seconds =
-                                                      value.inSeconds % 60;
-                                                  return Text(
-                                                    '$minutes:$seconds',
-                                                    style: TextStyle(
-                                                      color: UiConstants
-                                                          .primaryColor,
-                                                      fontWeight:
-                                                          FontWeight.bold,
-                                                    ),
-                                                  );
-                                                }),
-                                            // Text(
-                                            //   "15:00",
-                                            //   style: TextStyle(
-                                            //     color: UiConstants.primaryColor,
-                                            //     fontWeight: FontWeight.w700,
-                                            //   ),
-                                            // ),
-                                            Text("  minutes.")
-                                          ],
-                                        )
-                                      ],
-                                    ),
-                                  )
-                                : SizedBox(),
-                            _isOtpIncorrect
-                                ? Padding(
-                                    padding: const EdgeInsets.symmetric(
-                                        vertical: 16),
-                                    child: Row(
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.center,
-                                      children: [
-                                        Text(
-                                          "OTP is incorrect,please try again",
-                                          style: TextStyle(
-                                            color: Colors.red,
-                                            fontWeight: FontWeight.w700,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  )
-                                : SizedBox(),
-                            SizedBox(
-                              height: SizeConfig.screenHeight * 0.2,
-                            ),
-                          ],
+                          ),
                         ),
-                      ),
+                        Padding(
+                          padding: EdgeInsets.symmetric(
+                              vertical: SizeConfig.padding16),
+                          child: Text((!_isOtpSent)
+                              ? "We will send you a 6 digit code on this email."
+                              : "Please check your promotions or spam folders if you can't find the email"),
+                        ),
+                        SizedBox(
+                          height: SizeConfig.padding24,
+                        ),
+                        _isProcessing
+                            ? Container(
+                                width: SizeConfig.screenWidth,
+                                alignment: Alignment.center,
+                                child: Column(
+                                  children: [
+                                    CircularProgressIndicator(),
+                                    SizedBox(
+                                      height: 8,
+                                    ),
+                                    Text("Sending OTP")
+                                  ],
+                                ),
+                              )
+                            : SizedBox(),
+                        _isOtpSent
+                            ? Container(
+                                child: Column(
+                                  crossAxisAlignment: CrossAxisAlignment.start,
+                                  children: [
+                                    Text(
+                                      "Enter the OTP",
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.w700,
+                                          fontSize:
+                                              SizeConfig.cardTitleTextSize),
+                                    ),
+                                    SizedBox(
+                                      height: SizeConfig.padding16,
+                                    ),
+                                    Padding(
+                                      padding: const EdgeInsets.fromLTRB(
+                                          0, 18.0, 0, 18.0),
+                                      child: PinInputTextField(
+                                        autoFocus: true,
+                                        pinLength: 6,
+                                        decoration: BoxLooseDecoration(
+                                          enteredColor:
+                                              UiConstants.primaryColor,
+                                          solidColor: UiConstants.primaryColor
+                                              .withOpacity(0.04),
+                                          strokeColor: UiConstants.primaryColor,
+                                          strokeWidth: 1,
+                                          textStyle: GoogleFonts.montserrat(
+                                              fontSize: 20,
+                                              color: Colors.black),
+                                        ),
+                                        controller: otp,
+                                        onChanged: (value) {
+                                          print(value);
+                                          if (value.length == 6) {
+                                            verifyOtp();
+                                          } else {
+                                            setState(() {
+                                              _isOtpIncorrect = false;
+                                            });
+                                          }
+                                        },
+                                        onSubmit: (pin) {
+                                          print("Pressed submit for pin: " +
+                                              pin.toString() +
+                                              "\n  No action taken.");
+                                        },
+                                      ),
+                                    ),
+                                    SizedBox(
+                                      height: SizeConfig.padding16,
+                                    ),
+                                    Row(
+                                      children: [
+                                        Text("OTP is only valid for "),
+                                        TweenAnimationBuilder<Duration>(
+                                            duration: Duration(minutes: 15),
+                                            tween: Tween(
+                                                begin: Duration(minutes: 15),
+                                                end: Duration.zero),
+                                            onEnd: () {
+                                              print('Timer ended');
+                                              BaseUtil.showNegativeAlert(
+                                                "Session Expired!",
+                                                "Please try again",
+                                              );
+                                              AppState.backButtonDispatcher
+                                                  .didPopRoute();
+                                            },
+                                            builder: (BuildContext context,
+                                                Duration value, Widget child) {
+                                              final minutes = value.inMinutes;
+                                              final seconds =
+                                                  value.inSeconds % 60;
+                                              return Text(
+                                                '$minutes:$seconds',
+                                                style: TextStyle(
+                                                  color:
+                                                      UiConstants.primaryColor,
+                                                  fontWeight: FontWeight.bold,
+                                                ),
+                                              );
+                                            }),
+                                        // Text(
+                                        //   "15:00",
+                                        //   style: TextStyle(
+                                        //     color: UiConstants.primaryColor,
+                                        //     fontWeight: FontWeight.w700,
+                                        //   ),
+                                        // ),
+                                        Text("  minutes.")
+                                      ],
+                                    )
+                                  ],
+                                ),
+                              )
+                            : SizedBox(),
+                        _isOtpIncorrect
+                            ? Padding(
+                                padding:
+                                    const EdgeInsets.symmetric(vertical: 16),
+                                child: Row(
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  children: [
+                                    Text(
+                                      "OTP is incorrect,please try again",
+                                      style: TextStyle(
+                                        color: Colors.red,
+                                        fontWeight: FontWeight.w700,
+                                      ),
+                                    ),
+                                  ],
+                                ),
+                              )
+                            : SizedBox(),
+                        SizedBox(
+                          height: SizeConfig.screenHeight * 0.5,
+                        ),
+                      ],
                     ),
                     Align(
                       alignment: Alignment.bottomCenter,
