@@ -5,18 +5,15 @@ import 'package:felloapp/core/base_remote_config.dart';
 import 'package:felloapp/core/enums/page_state_enum.dart';
 import 'package:felloapp/core/model/base_user_model.dart';
 import 'package:felloapp/core/ops/https/http_ops.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
 import 'package:felloapp/core/service/fcm/fcm_listener_service.dart';
-import 'package:felloapp/core/service/mixpanel_service.dart';
 import 'package:felloapp/core/service/tambola_service.dart';
 import 'package:felloapp/core/service/user_service.dart';
 import 'package:felloapp/navigator/app_state.dart';
 import 'package:felloapp/navigator/router/ui_pages.dart';
 import 'package:felloapp/ui/architecture/base_vm.dart';
-import 'package:felloapp/ui/elements/logo/logo_canvas.dart';
-import 'package:felloapp/util/assets.dart';
 import 'package:felloapp/util/custom_logger.dart';
 import 'package:felloapp/util/locator.dart';
-import 'package:flutter/services.dart';
 import 'package:felloapp/util/custom_logger.dart';
 import 'package:package_info/package_info.dart';
 
@@ -33,7 +30,7 @@ class LauncherViewModel extends BaseModel {
   final _httpModel = locator<HttpModel>();
   final _logger = locator<CustomLogger>();
   final _tambolaService = locator<TambolaService>();
-  final _mixpanelService = locator<MixpanelService>();
+  final _analyticsService = locator<AnalyticsService>();
 
   //GETTERS
   bool get isSlowConnection => _isSlowConnection;
@@ -60,7 +57,7 @@ class LauncherViewModel extends BaseModel {
     await _baseUtil.init();
     _tambolaService.init();
     await _fcmListener.setupFcm();
-    await _mixpanelService.init(
+    await _analyticsService.login(
         isOnboarded: userService.isUserOnborded,
         baseUser: userService.baseUser);
     _httpModel.init();
@@ -126,7 +123,7 @@ class LauncherViewModel extends BaseModel {
     bool _res = false;
     try {
       _res = await deviceUnlock.request(
-          localizedReason: 'Please authenticate in order to proceed');
+          localizedReason: 'Confirm your phone screen lock pattern,PIN or password');
     } on DeviceUnlockUnavailable {
       BaseUtil.showPositiveAlert('No Device Authentication Found',
           'Logging in, please enable device security to add lock');
@@ -137,7 +134,7 @@ class LauncherViewModel extends BaseModel {
     } catch (e) {
       _logger.e("error", [e]);
       BaseUtil.showNegativeAlert('Authentication Failed',
-          'Please restart the application to try again.');
+          'Please restart and try again');
     }
     return _res;
   }
