@@ -29,114 +29,126 @@ class ReferralLeaderboard extends StatelessWidget {
             LeaderBoardServiceProperties>(
         properties: [LeaderBoardServiceProperties.ReferralLeaderboard],
         builder: (context, model, properties) {
-          return Column(
-            children: [
-              Container(
-                decoration: BoxDecoration(
-                  color: Colors.white,
+          return Container(
+            color: Colors.white,
+            child: Column(
+              children: [
+                Container(
+                  decoration: BoxDecoration(
+                    color: Colors.white,
+                  ),
+                  padding: EdgeInsets.only(top: SizeConfig.padding8),
+                  child: model.referralLeaderBoard == null
+                      ? Container(
+                          width: SizeConfig.screenWidth,
+                          color: Colors.white,
+                          alignment: Alignment.center,
+                          child: ListLoader())
+                      : (model.referralLeaderBoard.isEmpty
+                          ? Container(
+                              width: SizeConfig.screenWidth,
+                              height: SizeConfig.safeScreenHeight * 0.7,
+                              color: Colors.white,
+                              child: NoRecordDisplayWidget(
+                                asset: "images/leaderboard.png",
+                                text:
+                                    "Referral Leaderboard will be updated soon",
+                              ),
+                            )
+                          : Column(
+                              children: [
+                                Padding(
+                                  padding: EdgeInsets.symmetric(
+                                    horizontal:
+                                        SizeConfig.pageHorizontalMargins,
+                                  ),
+                                  child: Row(
+                                    mainAxisAlignment:
+                                        MainAxisAlignment.spaceBetween,
+                                    children: [
+                                      Text(
+                                        'This months\'s top referrers:',
+                                        style: TextStyles.body4
+                                            .colour(Colors.grey),
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                                Column(
+                                  children: List.generate(
+                                    getLength(model.referralLeaderBoard.length),
+                                    (i) {
+                                      return Container(
+                                        width: SizeConfig.screenWidth,
+                                        padding: EdgeInsets.all(
+                                            SizeConfig.padding12),
+                                        margin: EdgeInsets.symmetric(
+                                            vertical: SizeConfig.padding8,
+                                            horizontal: SizeConfig
+                                                .pageHorizontalMargins),
+                                        decoration: BoxDecoration(
+                                          color: Color(0xfff6f6f6),
+                                          borderRadius: BorderRadius.circular(
+                                              SizeConfig.roundness16),
+                                        ),
+                                        child: Row(
+                                          children: [
+                                            CircleAvatar(
+                                              backgroundColor:
+                                                  UiConstants.primaryColor,
+                                              radius: SizeConfig.padding16,
+                                              child: Text(
+                                                "${i + 1}",
+                                                style: TextStyles.body4
+                                                    .colour(Colors.white),
+                                              ),
+                                            ),
+                                            SizedBox(
+                                                width: SizeConfig.padding12),
+                                            Expanded(
+                                              child: Column(
+                                                crossAxisAlignment:
+                                                    CrossAxisAlignment.start,
+                                                mainAxisAlignment:
+                                                    MainAxisAlignment.center,
+                                                children: [
+                                                  Text(
+                                                      model
+                                                              .referralLeaderBoard[
+                                                                  i]
+                                                              .username
+                                                              .replaceAll(
+                                                                  '@', '.') ??
+                                                          "username",
+                                                      style: TextStyles.body3),
+                                                ],
+                                              ),
+                                            ),
+                                            PrizeChip(
+                                              color: UiConstants.primaryColor,
+                                              svg: Assets.plane,
+                                              text: model.referralLeaderBoard[i]
+                                                      .refCount
+                                                      .toString() ??
+                                                  "00",
+                                            ),
+                                          ],
+                                        ),
+                                      );
+                                    },
+                                  ),
+                                ),
+                              ],
+                            )),
                 ),
-                padding: EdgeInsets.only(top: SizeConfig.padding8),
-                child: model.referralLeaderBoard == null
-                    ? Container(
-                        width: SizeConfig.screenWidth,
-                        color: Colors.white,
-                        alignment: Alignment.center,
-                        child: ListLoader())
-                    : (model.referralLeaderBoard.isEmpty
-                        ? Container(
-                            width: SizeConfig.screenWidth,
-                            color: Colors.white,
-                            child: NoRecordDisplayWidget(
-                              asset: "images/leaderboard.png",
-                              text: "Referral Leaderboard will be updated soon",
-                            ),
-                          )
-                        : Column(
-                            children: [
-                              Padding(
-                                padding: EdgeInsets.symmetric(
-                                  horizontal: SizeConfig.pageHorizontalMargins,
-                                ),
-                                child: Row(
-                                  mainAxisAlignment:
-                                      MainAxisAlignment.spaceBetween,
-                                  children: [
-                                    Text(
-                                      'This months\'s top referrers:',
-                                      style:
-                                          TextStyles.body4.colour(Colors.grey),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                              Column(
-                                children: List.generate(
-                                  getLength(model.referralLeaderBoard.length),
-                                  (i) {
-                                    return Container(
-                                      width: SizeConfig.screenWidth,
-                                      padding:
-                                          EdgeInsets.all(SizeConfig.padding12),
-                                      margin: EdgeInsets.symmetric(
-                                          vertical: SizeConfig.padding8,
-                                          horizontal:
-                                              SizeConfig.pageHorizontalMargins),
-                                      decoration: BoxDecoration(
-                                        color: Color(0xfff6f6f6),
-                                        borderRadius: BorderRadius.circular(
-                                            SizeConfig.roundness16),
-                                      ),
-                                      child: Row(
-                                        children: [
-                                          CircleAvatar(
-                                            backgroundColor:
-                                                UiConstants.primaryColor,
-                                            radius: SizeConfig.padding16,
-                                            child: Text(
-                                              "${i + 1}",
-                                              style: TextStyles.body4
-                                                  .colour(Colors.white),
-                                            ),
-                                          ),
-                                          SizedBox(width: SizeConfig.padding12),
-                                          Expanded(
-                                            child: Column(
-                                              crossAxisAlignment:
-                                                  CrossAxisAlignment.start,
-                                              mainAxisAlignment:
-                                                  MainAxisAlignment.center,
-                                              children: [
-                                                Text(
-                                                    model.referralLeaderBoard[i]
-                                                            .username
-                                                            .replaceAll(
-                                                                '@', '.') ??
-                                                        "username",
-                                                    style: TextStyles.body3),
-                                              ],
-                                            ),
-                                          ),
-                                          PrizeChip(
-                                            color: UiConstants.primaryColor,
-                                            png: Assets.moneyIcon,
-                                            text: model.referralLeaderBoard[i]
-                                                    .refCount
-                                                    .toString() ??
-                                                "00",
-                                          ),
-                                        ],
-                                      ),
-                                    );
-                                  },
-                                ),
-                              ),
-                            ],
-                          )),
-              ),
-              SizedBox(
-                height: SizeConfig.navBarHeight * 1.5,
-              )
-            ],
+                SizedBox(
+                  height: model.referralLeaderBoard.length < 10
+                      ? (10 - model.referralLeaderBoard.length) *
+                          SizeConfig.padding54
+                      : SizeConfig.navBarHeight * 1.5,
+                )
+              ],
+            ),
           );
         });
   }
