@@ -1,13 +1,9 @@
 import 'package:felloapp/base_util.dart';
 import 'package:felloapp/core/base_remote_config.dart';
-import 'package:felloapp/core/enums/page_state_enum.dart';
 import 'package:felloapp/core/enums/user_service_enum.dart';
 import 'package:felloapp/core/model/tambola_winners_details.dart';
 import 'package:felloapp/core/service/user_service.dart';
-import 'package:felloapp/navigator/app_state.dart';
-import 'package:felloapp/navigator/router/ui_pages.dart';
 import 'package:felloapp/ui/pages/others/profile/my_winnings/my_winnings_vm.dart';
-import 'package:felloapp/ui/pages/others/profile/referrals/referral_history/referral_history_view.dart';
 import 'package:felloapp/ui/widgets/buttons/fello_button/fello_button.dart';
 import 'package:felloapp/util/assets.dart';
 import 'package:felloapp/util/styles/size_config.dart';
@@ -37,7 +33,7 @@ class PrizeClaimCard extends StatelessWidget {
                 if (m.userFundWallet.isPrizeBalanceUnclaimed())
                   Container(
                     margin: EdgeInsets.only(
-                        bottom: SizeConfig.padding24,
+                        top: SizeConfig.padding24,
                         left: SizeConfig.pageHorizontalMargins,
                         right: SizeConfig.pageHorizontalMargins),
                     decoration: BoxDecoration(
@@ -59,11 +55,9 @@ class PrizeClaimCard extends StatelessWidget {
                               CircleAvatar(
                                 backgroundColor:
                                     UiConstants.primaryColor.withOpacity(0.2),
-                                radius: SizeConfig.screenWidth * 0.067,
-                                child: Icon(
-                                  Icons.lock_open_sharp,
-                                  size: SizeConfig.padding32,
-                                  color: UiConstants.primaryColor,
+                                child: Image.asset(
+                                  Assets.moneyIcon,
+                                  width: SizeConfig.iconSize1,
                                 ),
                               ),
                               SizedBox(
@@ -76,22 +70,10 @@ class PrizeClaimCard extends StatelessWidget {
                                   children: [
                                     FittedBox(
                                       child: Text(
-                                        "Available Balance",
-                                        style: TextStyles.body1.bold,
+                                        "My Active Winnings: ",
+                                        style: TextStyles.body1,
                                       ),
                                     ),
-                                    SizedBox(height: SizeConfig.padding4),
-                                    (m.userFundWallet.unclaimedBalance >=
-                                                refUnlockAmt &&
-                                            m.userFundWallet.augGoldPrinciple >=
-                                                refUnlockAmt)
-                                        ? Text(
-                                            "Redeem your rewards using the following options",
-                                            maxLines: 2,
-                                            style: TextStyles.body3
-                                                .colour(Colors.grey),
-                                          )
-                                        : Container()
                                   ],
                                 ),
                               ),
@@ -105,7 +87,7 @@ class PrizeClaimCard extends StatelessWidget {
                           ),
                         ),
                         SizedBox(height: SizeConfig.padding8),
-                        if (m.userFundWallet.unclaimedBalance >= refUnlockAmt &&
+                        if (m.userFundWallet.unclaimedBalance >= minWithdrawPrizeAmt &&
                             m.userFundWallet.augGoldPrinciple >= refUnlockAmt)
                           Container(
                             margin: EdgeInsets.symmetric(
@@ -126,7 +108,7 @@ class PrizeClaimCard extends StatelessWidget {
                                     : SizedBox(),
                                 ClaimButton(
                                   color: UiConstants.tertiarySolid,
-                                  image: Assets.goldClaim,
+                                  image: "images/augmont-share.png",
                                   onTap: () => model.showConfirmDialog(
                                       PrizeClaimChoice.GOLD_CREDIT),
                                   text: "Redeem as Digital Gold",
@@ -134,7 +116,7 @@ class PrizeClaimCard extends StatelessWidget {
                               ],
                             ),
                           ),
-                        if (m.userFundWallet.unclaimedBalance < 100)
+                        if (m.userFundWallet.unclaimedBalance < minWithdrawPrizeAmt)
                           Container(
                             margin: EdgeInsets.symmetric(
                                 vertical: SizeConfig.padding6),
@@ -180,78 +162,6 @@ class PrizeClaimCard extends StatelessWidget {
                       ],
                     ),
                   ),
-                if (m.userFundWallet.lockedPrizeBalance > 0)
-                  InkWell(
-                    onTap: () {
-                      AppState.delegate.appState.currentAction = PageAction(
-                        state: PageState.addWidget,
-                        page: ReferralHistoryPageConfig,
-                        widget: ReferralHistoryView(
-                          onlyLocked: true,
-                        ),
-                      );
-                    },
-                    child: Container(
-                      margin: EdgeInsets.only(
-                        left: SizeConfig.pageHorizontalMargins,
-                        right: SizeConfig.pageHorizontalMargins,
-                        bottom: SizeConfig.padding24,
-                      ),
-                      //height: SizeConfig.screenWidth * 0.28,
-                      decoration: BoxDecoration(
-                        color: UiConstants.tertiaryLight,
-                        borderRadius:
-                            BorderRadius.circular(SizeConfig.roundness16),
-                      ),
-                      padding: EdgeInsets.symmetric(
-                        horizontal: SizeConfig.padding16,
-                        vertical: SizeConfig.padding20,
-                      ),
-                      child: Row(
-                        children: [
-                          CircleAvatar(
-                            backgroundColor:
-                                UiConstants.tertiarySolid.withOpacity(0.2),
-                            radius: SizeConfig.screenWidth * 0.067,
-                            child: Icon(
-                              Icons.lock_outline,
-                              size: SizeConfig.padding32,
-                              color: UiConstants.tertiarySolid,
-                            ),
-                          ),
-                          SizedBox(
-                            width: SizeConfig.padding12,
-                          ),
-                          Expanded(
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                FittedBox(
-                                  child: Text(
-                                    "Locked Balance",
-                                    style: TextStyles.body1.bold,
-                                  ),
-                                ),
-                                SizedBox(height: SizeConfig.padding4),
-                                Text(
-                                  "Unlock these rewards once your referred friends make their first savings of atleast ₹$refUnlock",
-                                  maxLines: 3,
-                                  style: TextStyles.body3.colour(Colors.grey),
-                                )
-                              ],
-                            ),
-                          ),
-                          SizedBox(width: SizeConfig.padding12),
-                          Text(
-                            "₹${m.userFundWallet.lockedPrizeBalance ?? '-'}",
-                            style: TextStyles.body1.bold
-                                .colour(UiConstants.tertiarySolid),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
               ],
             ));
   }
@@ -295,21 +205,29 @@ class ClaimButton extends StatelessWidget {
             vertical: SizeConfig.padding16,
           ),
           child: Row(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            mainAxisAlignment: MainAxisAlignment.center,
             children: [
               CircleAvatar(
                 radius: SizeConfig.padding24,
                 backgroundColor: Colors.white.withOpacity(0.3),
-                child: Image.asset(
-                  image ?? Assets.amazonClaim,
+                child: Padding(
+                  padding: EdgeInsets.all(SizeConfig.padding6),
+                  child: Image.asset(
+                    image ?? Assets.amazonClaim,
+                  ),
                 ),
               ),
-              SizedBox(width: SizeConfig.padding6),
-              Expanded(
+              SizedBox(width: SizeConfig.padding12),
+              Container(
+                width: SizeConfig.screenWidth * 0.5,
                 child: Text(
                   text ?? "Redeem for amazon pay",
-                  style: TextStyles.body3.colour(Colors.white),
+                  style: TextStyles.body1.colour(Colors.white).bold,
+                  maxLines: 2,
+                  textAlign: TextAlign.start,
                 ),
-              )
+              ),
             ],
           ),
         ),
