@@ -4,6 +4,7 @@ import 'package:felloapp/navigator/app_state.dart';
 import 'package:felloapp/navigator/router/ui_pages.dart';
 import 'package:felloapp/ui/architecture/base_view.dart';
 import 'package:felloapp/ui/pages/others/profile/userProfile/userProfile_view.dart';
+import 'package:felloapp/ui/pages/root/root_vm.dart';
 import 'package:felloapp/ui/service_elements/user_service/profile_image.dart';
 import 'package:felloapp/ui/service_elements/user_service/user_name_text.dart';
 import 'package:felloapp/ui/widgets/drawer/drawer_vm.dart';
@@ -30,6 +31,10 @@ class FDrawer extends StatelessWidget {
                   children: [
                     InkWell(
                         onTap: () {
+                          if (RootViewModel
+                              .scaffoldKey.currentState.isDrawerOpen)
+                            RootViewModel.scaffoldKey.currentState
+                                .openEndDrawer();
                           AppState.delegate.appState.currentAction = PageAction(
                             state: PageState.addWidget,
                             page: UserProfileDetailsConfig,
@@ -86,10 +91,7 @@ class FDrawer extends StatelessWidget {
                                 vertical: SizeConfig.padding8),
                             child: TextButton.icon(
                               onPressed: () {
-                                AppState.delegate.appState.currentAction =
-                                    PageAction(
-                                        state: PageState.addPage,
-                                        page: model.drawerList[i].pageConfig);
+                                model.onItemSelected(i);
                               },
                               icon: SvgPicture.asset(
                                 model.drawerList[i].icon,
