@@ -97,7 +97,7 @@ class MessageLookup extends MessageLookupByLibrary {
     "pkStateHint" : MessageLookupByLibrary.simpleMessage("Which state do you live in?"),
     "pkStateLabel" : MessageLookupByLibrary.simpleMessage("State"),
     "playTrendingGames" : MessageLookupByLibrary.simpleMessage("Trending Games"),
-    "refHIW" : MessageLookupByLibrary.simpleMessage("How it works?"),
+    "refHIW" : MessageLookupByLibrary.simpleMessage("How do you earn?"),
     "refShareLink" : MessageLookupByLibrary.simpleMessage("Share Link"),
     "refStep2" : MessageLookupByLibrary.simpleMessage("Your friend makes their first saving of ₹ 100 on the app."),
     "refStep3" : MessageLookupByLibrary.simpleMessage("Both you and your friend receive ₹ 25 and 200 Fello tokens in your account."),
@@ -119,7 +119,7 @@ class MessageLookup extends MessageLookupByLibrary {
     "saveSecure" : MessageLookupByLibrary.simpleMessage("100% secure"),
     "saveSellButton" : MessageLookupByLibrary.simpleMessage("SELL"),
     "saveViewAll" : MessageLookupByLibrary.simpleMessage("View All"),
-    "saveWinningsLabel" : MessageLookupByLibrary.simpleMessage("My Total Winnings"),
+    "saveWinningsLabel" : MessageLookupByLibrary.simpleMessage("My Active Winnings"),
     "saveWinningsValue" : m6,
     "signout" : MessageLookupByLibrary.simpleMessage("Sign Out"),
     "splashNoInternet" : MessageLookupByLibrary.simpleMessage("No active internet connection"),
@@ -138,7 +138,7 @@ class MessageLookup extends MessageLookupByLibrary {
     "tWinTitle" : MessageLookupByLibrary.simpleMessage("CONGRATULATIONS!"),
     "txnBankDetailsLabel" : MessageLookupByLibrary.simpleMessage("Bank Account Details"),
     "txnHistoryLabel" : MessageLookupByLibrary.simpleMessage("Transaction History"),
-    "winIphoneBigText" : MessageLookupByLibrary.simpleMessage("friends and win iphone 13"),
+    "winIphoneBigText" : MessageLookupByLibrary.simpleMessage("friends and win iPhone 13"),
     "winIphoneSmallText" : MessageLookupByLibrary.simpleMessage("Refer"),
     "winMoneyBigText" : MessageLookupByLibrary.simpleMessage("₹ 1 Lakh every week"),
     "winMoneySmallText" : MessageLookupByLibrary.simpleMessage("Play and win"),
