@@ -14,13 +14,16 @@ import 'package:felloapp/core/repository/statistics_repo.dart';
 import 'package:felloapp/core/repository/ticket_generation_repo.dart';
 import 'package:felloapp/core/repository/user_repo.dart';
 import 'package:felloapp/core/repository/winners_repo.dart';
+import 'package:felloapp/core/service/analytics/analytics_service.dart';
+import 'package:felloapp/core/service/analytics/mixpanel_analytics.dart';
 import 'package:felloapp/core/service/api.dart';
 import 'package:felloapp/core/service/connectivity_service.dart';
 import 'package:felloapp/core/service/fcm/fcm_handler_service.dart';
 import 'package:felloapp/core/service/fcm/fcm_listener_service.dart';
+import 'package:felloapp/core/service/golden_ticket_service.dart';
 import 'package:felloapp/core/service/lcl_db_api.dart';
 import 'package:felloapp/core/service/leaderboard_service.dart';
-import 'package:felloapp/core/service/mixpanel_service.dart';
+import 'package:felloapp/core/service/analytics/webengage_analytics.dart';
 import 'package:felloapp/core/service/payment_service.dart';
 import 'package:felloapp/core/service/prize_service.dart';
 import 'package:felloapp/core/service/tambola_service.dart';
@@ -48,6 +51,12 @@ import 'package:felloapp/ui/pages/others/profile/my_winnings/my_winnings_vm.dart
 import 'package:felloapp/ui/pages/others/profile/referrals/referral_details/referral_details_vm.dart';
 import 'package:felloapp/ui/pages/others/profile/transactions_history/transaction_history_vm.dart';
 import 'package:felloapp/ui/pages/others/profile/userProfile/userProfile_viewModel.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_milestones/golden_milestones_view.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_milestones/golden_milestones_vm.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_scratch_card/gt_detailed_vm.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_scratch_dialog/gt_instant_view.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_scratch_dialog/gt_instant_vm.dart';
+import 'package:felloapp/ui/pages/others/rewards/golden_tickets/golden_tickets_vm.dart';
 import 'package:felloapp/ui/pages/root/root_vm.dart';
 import 'package:felloapp/ui/pages/splash/splash_vm.dart';
 import 'package:felloapp/ui/widgets/coin_bar/coin_bar_vm.dart';
@@ -72,9 +81,12 @@ void setupLocator() {
   locator.registerLazySingleton(() => FcmListener());
   locator.registerLazySingleton(() => FcmHandler());
 
+  locator.registerLazySingleton(() => AnalyticsService());
+  locator.registerLazySingleton(() => MixpanelAnalytics());
+  locator.registerLazySingleton(() => WebEngageAnalytics());
+
   //Model Services
   locator.registerLazySingleton(() => BaseUtil());
-  locator.registerLazySingleton(() => MixpanelService());
   locator.registerLazySingleton(() => PaymentService());
   locator.registerLazySingleton(() => AppState());
   locator.registerLazySingleton(() => ConnectivityService());
@@ -85,6 +97,7 @@ void setupLocator() {
   locator.registerLazySingleton(() => PrizeService());
   locator.registerLazySingleton(() => WinnerService());
   locator.registerLazySingleton(() => LeaderboardService());
+  locator.registerLazySingleton(() => GoldenTicketService());
 
   //Repository
   locator.registerLazySingleton(() => DBModel());
@@ -128,6 +141,10 @@ void setupLocator() {
   locator.registerFactory(() => ReferralDetailsViewModel());
   locator.registerFactory(() => MyWinningsViewModel());
   locator.registerFactory(() => NotificationsViewModel());
+  locator.registerFactory(() => GoldenTicketsViewModel());
+  locator.registerFactory(() => GTDetailedViewModel());
+  locator.registerFactory(() => GoldenMilestonesViewModel());
+  locator.registerFactory(() => GTInstantViewModel());
 
   //WIDGETS
   // locator.registerFactory(() => SellGoldBtnVM());
@@ -137,5 +154,4 @@ void setupLocator() {
   locator.registerFactory(() => FAQCardViewModel());
 
   //....
-
 }
